Changed Datatype

+++ reservoir_management/reservoir/views.py
@@ -142,12 +142,12 @@ def calculate_reservoir_health_score(request):
     Returns:
     - float: Final reservoir health score (0-100).
     """
-    storage_tmc = int(request.GET.get("current_storage"))
-    capacity_tmc = int(request.GET.get("gross_capacity"))
-    siltation_tmc = int(request.GET.get("siltation"))
-    flood_cushion_tmc = int(request.GET.get("flood_cushion"))
-    evaporation_mm = int(request.GET.get("evaporation"))
-    rainfall_mm = int(request.GET.get("rainfall"))
+    storage_tmc = float(request.GET.get("current_storage"))
+    capacity_tmc = float(request.GET.get("gross_capacity"))
+    siltation_tmc = float(request.GET.get("siltation"))
+    flood_cushion_tmc = float(request.GET.get("flood_cushion"))
+    evaporation_mm = float(request.GET.get("evaporation"))
+    rainfall_mm = float(request.GET.get("rainfall"))
     age_years = int(request.GET.get("age"))
     design_life_years = int(request.GET.get("design_life"))
